Count should be always integer

+++ tests/lib/Absolute.php
@@ -33,7 +33,7 @@ trait Absolute
     public function initCounts(App $app, array $_values) {
         $initCount    = count($_values[Layout\Collection::INITIAL_VALUES][0]);
         $teacherCount = count($_values[Layout\Collection::TEACHER_VALUES][0]);
-        $assocCount   = $initCount * NeuronNet\Relation::ACCELERATOR; //The number of associated neurons should be smaller than the number of stimulated
+        $assocCount   = (int)($initCount * NeuronNet\Relation::ACCELERATOR); //The number of associated neurons should be smaller than the number of stimulated
         //Setting counts
         $app->setNeuronCount(NeuronNet\Factory::STIM, $initCount);
         $app->setNeuronCount(NeuronNet\Factory::ASSOC, $assocCount);
